feat: Split extended info section to own component + Style extended info section as in figma
